pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "xnat-shiny-uploader"
version = "0.1.0"
description = "Bulk upload NIfTI + labels to XNAT subjects (plus optional demographics)."
requires-python = ">=3.9"
dependencies = ["requests>=2.31.0", "PyYAML>=6.0.1"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
